Refactor triangle
